Se agrega solucion de evitar que cualquiera pueda cancelar un turno

+++ Odontologia/Doctores/views.py
@@ -145,14 +145,22 @@ def cancelando_turno(request):
     if request.method == "POST":
         formu = forms.CancelandoTurnos(request.POST)
         if formu.is_valid():
-            conex = sqlite3.connect("db.sqlite3")
-            cursor = conex.cursor()
-            id_turno_actual = formu.cleaned_data['ID_Turno_actual']
-            sql = ('DELETE FROM Doctores_turnosasignados WHERE ID_Turno_Disponible = ?', (id_turno_actual,))
-            cursor.execute(*sql)
-            conex.commit()
-            conex.close()
-            return render(request, 'turno_cancelado.html')
+            id_formulario = formu.cleaned_data['ID_Turno_actual']
+            dni_formulario = formu.cleaned_data['DNI']
+            filtro = TurnosAsignados.objects.get(ID_Turno_Disponible = id_formulario)
+            filtro_dni = filtro.DNI
+            if filtro_dni == dni_formulario:
+                conex = sqlite3.connect("db.sqlite3")
+                cursor = conex.cursor()
+                id_turno_actual = formu.cleaned_data['ID_Turno_actual']
+                sql = ('DELETE FROM Doctores_turnosasignados WHERE ID_Turno_Disponible = ?', (id_turno_actual,))
+                cursor.execute(*sql)
+                conex.commit()
+                conex.close()
+                return render(request, 'turno_cancelado.html')
+
+            else:
+                return render(request, 'error_cancelar.html')
 
     else:
         formu = forms.CancelandoTurnos()
